improve driver detection
determineDriver() is now data driven. The decision about which driver
to use is pushed to the drivers themselves. The first matching driver is
what we'll use. It improves the detection by not using uname for BB, which
has many edge cases. Also makes it easier to add new drivers in the future
without resorting to string matching hacks.